QUAM/backend/requirements.txt
annotated-types==0.7.0
anyio>=4.0.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.0
filelock==3.16.1
fsspec==2024.6.1
git-filter-repo==2.47.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20250625
pydantic==2.10.6
pydantic_core==2.27.2
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tiktoken==0.7.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
yt-dlp==2024.10.22
python-multipart
